Details pannel
Details pannel appears on marker click (css transition).
It displays the rating, title and space of a location

We use Handlebars for displaying the templates. We added a transform method to
browserify in order to handle the compilation.
The npm package https://github.com/dlmanning/browserify-handlebars should be
updated, since it uses an outdated version of Handlebars. Also, a system has to
be defined in order to include Handlebars helpers.